+++ src/i18n/locales/ko.json
@@ -122,5 +122,10 @@
   "requireAuth": {
     "cannotAccess": "이 주소로 접근하기 위해서는 로그인이 필요합니다.",
     "goToLogIn": "로그인 페이지로 이동 →"
+  },
+  "invalidId": {
+    "reservedWords": "admin, user 등의 단어는 사용할 수 없습니다.",
+    "invalidCharacters": "영어 소문자, 숫자, 하이픈(-), 언더스코어(_)만 사용할 수 있습니다.",
+    "lengthError": "아이디는 3자 이상 20자 이하로 입력해야 합니다."
   }
 }
